StreamEdgeCore/connector/requirements.txt
zmq
paho-mqtt
pyyaml
